fixing up styling and move EventSummary components into their own file #3 #2 #10
