+++ src/EventList.css
@@ -1,6 +1,22 @@
 .event-box {
     padding: 0px;
+    min-height: 75px;
     display: flex;
+}
+
+.datetime-box{
+    height: 100%;
+}
+
+.timestamp{
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.type-box{
+    height: 100%;
 }
 
 .type-icon {
@@ -9,4 +25,11 @@
 
 .event-summary {
     padding-bottom: 5px;
+}
+
+.rawPayloadData-box{
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center; 
 }
